requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
requests==2.32.3

pytest==8.2.0
playwright==1.47.0

# YAML / tooling
pyyaml==6.0.2

# Lint
ruff==0.6.9

# Selenium + Visual AI
selenium>=4.20,<5
eyes-selenium>=5,<6
